requirements.txt
Flask==2.3.2
moviepy==2.1.3
Pillow==10.0.1
gTTS==2.4.0
requests==2.31.0
python-dotenv==1.0.0
mutagen==1.46.0
gunicorn==20.1.0
